add create database

+++ db/tables.py
@@ -6,4 +6,4 @@ class Products(Base):
     __tablename__ = "Products"
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String)
-    price = Column(DOUBLE)
+    price = Column(Integer)
